Day 4: Count Objects
Complete the function in the editor. It has one parameter: an array, a, of objects. Each object in the array has two integer properties denoted by x and y. The function must return a count of all such objects o in array a that satisfy o.x = o.y.
Sample Input 
5
1 1
2 3
3 3
3 4
4 5
Sample Output 
2