Day 5: Inheritance
We provide the implementation for a Rectangle class in the editor. Perform the following tasks:
Add an area method to Rectangle’s prototype.
Create a Square class that satisfies the following:
It is a subclass of Rectangle.
It contains a constructor and no other methods.
It can use the Rectangle class’ area method to print the area of a Square object.